[Instructions] Load | Store -> no more memory leak

+++ Sources/Factory.cpp
@@ -48,20 +48,20 @@ namespace Abstract
     {
         if (std::stof(value) > std::numeric_limits<float>::max() || std::stof(value) < std::numeric_limits<float>::min())
             throw Exception("Type value overflow");
-        return new Operand<float>(std::stof(value), Float);
+        return new Operand<float>(std::stold(value), Float);
     }
 
     IOperand* Factory::createDouble(const std::string &value)
     {
         if (std::stod(value) > std::numeric_limits<double>::max() || std::stod(value) < std::numeric_limits<double>::min())
             throw Exception("Type value overflow");
-        return new Operand<double>(std::stof(value), Double);
+        return new Operand<double>(std::stold(value), Double);
     }
 
     IOperand* Factory::createBigDecimal(const std::string &value)
     {
         if (std::stod(value) > std::numeric_limits<long double>::max() || std::stod(value) < std::numeric_limits<long double>::min())
             throw Exception("Type value overflow");
-        return new Operand<long double>(std::stof(value), BigDecimal);
+        return new Operand<long double>(std::stold(value), BigDecimal);
     }
 }
